conformal_energy_forecasting/requirements.txt
torch==2.1.0
numpy<2
mlflow==2.9.2
azureml-mlflow
azureml-core
azureml-dataset-runtime[fuse]
azure-ai-ml
pandas
